Fixed __eq__ and __hash__ functions in BLEUUID. (#152)
When full UUID object (with UUIDBase) is used to write to a
characteristic (eg. Nordic UART), the returned UUID object from API (in
on_notification) is in shortened form, and does not contain the full
base UUID. However, it can safely be compared just by the 'type' field
and value.

Similarly, these objects could not be hashed, because self.base.base is
None, not list, and thus the __hash__ function fails inside
sum(self.base.base).